allow for larger smooth group values
git-svn-id: http://ardorlabs.svn.cvsdude.com/Ardor3Dv1/trunk@1384 82abaa13-a93f-4eff-8987-b3ae9fc18050

+++ ardor3d-core/src/main/java/com/ardor3d/util/geom/GeometryTool.java
@@ -112,7 +112,7 @@ public abstract class GeometryTool {
             final Map<VertKey, Integer> store = Maps.newHashMap();
             final Map<Integer, Integer> indexRemap = Maps.newHashMap();
             int good = 0;
-            int group;
+            long group;
             for (int x = 0, max = verts.length; x < max; x++) {
                 group = groupData.getGroupForVertex(x);
                 final VertKey vkey = new VertKey(verts[x], norms != null ? norms[x] : null, colors != null ? colors[x]
